+++ src/motor_control.cpp
@@ -1,0 +1,63 @@
+#include "motor_control.h"
+#include "constants.h"
+
+namespace {
+const int veloInicial = 60;     // Velocidad mínima definida por el usuario
+const int veloMax = 170;        // Velocidad máxima
+int velo = veloInicial;         // Velocidad actual
+}
+
+void initMotors() {
+    pinMode(PWMA, OUTPUT);
+    pinMode(PWMB, OUTPUT);
+    pinMode(AIN1, OUTPUT);
+    pinMode(AIN2, OUTPUT);
+    pinMode(BIN1, OUTPUT);
+    pinMode(BIN2, OUTPUT);
+}
+
+void motorA_forward(int speed) {
+    digitalWrite(AIN1, LOW);
+    digitalWrite(AIN2, HIGH);
+    analogWrite(PWMA, speed);
+}
+
+void motorA_backward(int speed) {
+    digitalWrite(AIN1, HIGH);
+    digitalWrite(AIN2, LOW);
+    analogWrite(PWMA, speed);
+}
+
+void motorB_forward(int speed) {
+    digitalWrite(BIN1, HIGH);
+    digitalWrite(BIN2, LOW);
+    analogWrite(PWMB, speed);
+}
+
+void motorB_backward(int speed) {
+    digitalWrite(BIN1, LOW);
+    digitalWrite(BIN2, HIGH);
+    analogWrite(PWMB, speed);
+}
+
+void stopMotors() {
+    digitalWrite(AIN1, LOW);
+    digitalWrite(AIN2, LOW);
+    digitalWrite(BIN1, LOW);
+    digitalWrite(BIN2, LOW);
+    analogWrite(PWMA, 0);
+    analogWrite(PWMB, 0);
+}
+
+void setSpeed(char cmd) {
+    if (cmd >= '0' && cmd <= '9') {
+        int porcentaje = (cmd - '0') * 10; // 0%, 10%, ..., 90%
+        velo = veloInicial + ((veloMax - veloInicial) * porcentaje) / 100;
+    } else if (cmd == 'q') {
+        velo = veloMax;
+    }
+}
+
+int getSpeed() {
+    return velo;
+}
